+++ log.go
@@ -1,32 +1,34 @@
 package main
 
 import (
-	"github.com/gookit/color"
+	"fmt"
+	"github.com/TwinProduction/go-color"
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"runtime"
 	"strings"
 	"time"
 )
 
 func LogInfo(reference, data string) {
-	color.Green.Println(time.Now().Format("2006-01-02 15:04:05.000") + " [" + reference + "] --INF-- " + data)
+	fmt.Println(color.Ize(color.Green, time.Now().Format("2006-01-02 15:04:05.000")+" ["+reference+"] --INF-- "+data))
 	AppendDataToLog("INF", reference, data)
 }
 
 func LogError(reference, data string) {
-	color.Red.Println(time.Now().Format("2006-01-02 15:04:05.000") + " [" + reference + "] --ERR-- " + data)
+	fmt.Println(color.Ize(color.Red, time.Now().Format("2006-01-02 15:04:05.000")+" ["+reference+"] --INF-- "+data))
 	AppendDataToLog("ERR", reference, data)
 	AppendDataToErrLog("ERR", reference, data)
 }
 
 func LogWarning(reference, data string) {
-	color.Yellow.Println(time.Now().Format("2006-01-02 15:04:05.000") + " [" + reference + "] --WAR-- " + data)
+	fmt.Println(color.Ize(color.Yellow, time.Now().Format("2006-01-02 15:04:05.000")+" ["+reference+"] --INF-- "+data))
 	AppendDataToLog("WRN", reference, data)
 }
 
 func LogDebug(reference, data string) {
-	color.Blue.Println(time.Now().Format("2006-01-02 15:04:05.000") + " [" + reference + "] --DEB-- " + data)
+	fmt.Println(color.Ize(color.Blue, time.Now().Format("2006-01-02 15:04:05.000")+" ["+reference+"] --INF-- "+data))
 }
 
 func LogDirectoryFileCheck(reference string) {
@@ -35,13 +37,13 @@ func LogDirectoryFileCheck(reference string) {
 	_, checkPathError := os.Stat(logDirectory)
 	logDirectoryExists := checkPathError == nil
 	if logDirectoryExists {
-		color.Blue.Println(time.Now().Format(dateTimeFormat) + " [" + reference + "] --DEB-- " + "Log directory already exists ")
+		fmt.Println(color.Ize(color.Blue, time.Now().Format(dateTimeFormat)+" ["+reference+"] --DEB-- "+"Log directory already exists "))
 		return
 	}
-	color.Yellow.Println(time.Now().Format(dateTimeFormat) + " [" + reference + "] --WRN-- " + "Log directory does not exist, creating")
+	fmt.Println(color.Ize(color.Yellow, time.Now().Format(dateTimeFormat)+" ["+reference+"] --WRN-- "+"Log directory does not exist, creating"))
 	mkdirError := os.MkdirAll(logDirectory, 0777)
 	if mkdirError != nil {
-		color.Red.Println(time.Now().Format(dateTimeFormat) + " [" + reference + "] --ERR--" + "Unable to create directory for log file: " + mkdirError.Error())
+		fmt.Println(color.Ize(color.Red, time.Now().Format(dateTimeFormat)+" ["+reference+"] --ERR--"+"Unable to create directory for log file: "+mkdirError.Error()))
 		return
 	}
 }
@@ -54,13 +56,13 @@ func AppendDataToLog(logLevel string, reference string, data string) {
 	logFullPath := strings.Join([]string{logDirectory, logFileName}, "/")
 	f, err := os.OpenFile(logFullPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		color.Red.Println(time.Now().Format(dateTimeFormat) + " [" + reference + "] --WAR-- " + "Log file not present: " + err.Error())
+		fmt.Println(color.Ize(color.Yellow, time.Now().Format(dateTimeFormat)+" ["+reference+"] --WAR-- "+"Log file not present: "+err.Error()))
 		return
 	}
 	defer f.Close()
 	logData := time.Now().Format("2006-01-02 15:04:05.000   ") + reference + "   " + logLevel + "   " + data
 	if _, err := f.WriteString(logData + "\r\n"); err != nil {
-		color.Red.Println(time.Now().Format(dateTimeFormat) + " [" + reference + "] --ERR-- " + "Cannot write to file: " + err.Error())
+		fmt.Println(color.Ize(color.Red, time.Now().Format(dateTimeFormat)+" ["+reference+"] --ERR-- "+"Cannot write to file: "+err.Error()))
 	}
 }
 
@@ -72,17 +74,19 @@ func AppendDataToErrLog(logLevel string, reference string, data string) {
 	logFullPath := strings.Join([]string{logDirectory, logFileName}, "/")
 	f, err := os.OpenFile(logFullPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		color.Red.Println(time.Now().Format(dateTimeFormat) + " [" + reference + "] --WAR-- " + "Log file not present: " + err.Error())
+		fmt.Println(color.Ize(color.Red, time.Now().Format(dateTimeFormat)+" ["+reference+"] --WAR-- "+"Log file not present: "+err.Error()))
 		return
 	}
 	defer f.Close()
 	logData := time.Now().Format("2006-01-02 15:04:05.000   ") + reference + "   " + logLevel + "   " + data
 	if _, err := f.WriteString(logData + "\r\n"); err != nil {
-		color.Red.Println(time.Now().Format(dateTimeFormat) + " [" + reference + "] --ERR-- " + "Cannot write to file: " + err.Error())
+		fmt.Println(color.Ize(color.Red, time.Now().Format(dateTimeFormat)+" ["+reference+"] --ERR-- "+"Cannot write to file: "+err.Error()))
 	}
 }
 
 func DeleteOldLogFiles() {
+	LogInfo("MAIN", "Deleting old log files")
+	timer := time.Now()
 	directory, err := ioutil.ReadDir("log")
 	if err != nil {
 		LogError("MAIN", "Problem opening log directory")
@@ -100,4 +104,19 @@ func DeleteOldLogFiles() {
 			}
 		}
 	}
+	LogInfo("MAIN", "Old log files deleted, elapsed: "+time.Since(timer).String())
+}
+
+func GetDirectory() string {
+	var dir string
+	if runtime.GOOS == "windows" {
+		executable, err := os.Executable()
+		if err != nil {
+			panic(err)
+		}
+		dir = filepath.Dir(executable)
+	} else {
+		dir, _ = os.Getwd()
+	}
+	return dir
 }